requirements.txt
python-frontmatter
nsapiwrapper
configparser
guietta
